# Optimized packet load in gemm_pack_rhs (for Packet16f and Packet8f)