@anuj-rawat 2019-05-01
Hi Eugene. I have now made the following changes to the patch:

1.  Removed `MaskType` from template parameter list for `OptimizedPacketLoadOverTwoColumns`. Instead, `unpacket_traits<PacketType>::mask_t` is used. This also removes the `unsigned long long` from the `enable_if` later.
2. Minor refactoring of `loadPartialPacketStandard`, `loadPacketStandardFromTwoColumns`, `loadPacketStandardFromSingleColumn` and `loadPacketStandard` for (hopefully) improved readability. This includes:
  a. Renamed variables to drop 'Part' suffix.
  b. Reuse variable names as much as possible esp. where they "mean" the same for partial packet and packet.
  c. Added some comments explaining the intended behavior of code blocks.